requirements.txt
Twisted~=25.5.0
gentle~=0.11.0
setuptools~=80.9.0
openai-whisper~=20250625
requests~=2.32.4
torch~=2.8.0